feat: add publisher name and namespace dimensions
By using a shared map between the ebpf programs, store node information
when it is created using the pid as key, to recognize the same pid
sending data to the ROS topic to extrapolate the info.
Modified the Makefile to build the go program and delete leftovers
from the C userspace program.

fixes #1