Publish metric about messages taken from a topic
Hooking a tracepoint into the rmw_take_with_info function, publish a
counter about how many messages are taken by a given subscriber node for
a topic. If the number starts diverging from the one of messages
published on the topic, than we might be taking too long to process
messages taken from the callback subscription.